Add a wait loop to wait for the MySQL server to start
The mysql_install_db script that runs the first time you run the mysql
container is pretty slow (10 seconds or so), so if you were too fast
in starting this container (e.g. if you tried starting all containers
using Fig) it would fail to connect to the MySQL server.

Thus we have to retry until MySQL is up, by borrowing a wait loop from
tutum-docker-mysql.